+++ test/intl_phone_field_test.dart
@@ -1,9 +1,10 @@
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:intl_phone_field/intl_phone_field.dart';
 
 class TestWidget extends StatelessWidget {
-  const TestWidget({Key? key, required this.phoneNumber, this.countryCode}) : super(key: key);
+  const TestWidget({Key? key, required this.phoneNumber, this.countryCode})
+    : super(key: key);
 
   final String phoneNumber;
   final String? countryCode;
@@ -11,22 +12,23 @@ class TestWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-        title: 'Test Intl Phone Field',
-        home: Scaffold(
-          appBar: AppBar(title: const Text("")),
-          body: IntlPhoneField(
-            initialValue: phoneNumber,
-            initialCountryCode: countryCode,
-          ),
-        ));
+      title: 'Test Intl Phone Field',
+      home: Scaffold(
+        appBar: AppBar(title: const Text("")),
+        body: IntlPhoneField(
+          initialValue: phoneNumber,
+          initialCountryCode: countryCode,
+        ),
+      ),
+    );
   }
 }
 
 void main() {
-  testWidgets('Test intl_phone_field setup with completeNumber', (WidgetTester tester) async {
-    await tester.pumpWidget(const TestWidget(
-      phoneNumber: '+447891234467',
-    ));
+  testWidgets('Test intl_phone_field setup with completeNumber', (
+    WidgetTester tester,
+  ) async {
+    await tester.pumpWidget(const TestWidget(phoneNumber: '+447891234467'));
 
     final countryCodeFinder = find.text('+44');
     final numberFinder = find.text('7891234467');
@@ -35,24 +37,27 @@ void main() {
     expect(numberFinder, findsOneWidget);
   });
 
-  testWidgets('Test intl_phone_field setup with Guernsey number: +441481960194', (WidgetTester tester) async {
-    await tester.pumpWidget(const TestWidget(
-      phoneNumber: '+441481960194',
-      countryCode: 'GG',
-    ));
+  testWidgets(
+    'Test intl_phone_field setup with Guernsey number: +441481960194',
+    (WidgetTester tester) async {
+      await tester.pumpWidget(
+        const TestWidget(phoneNumber: '+441481960194', countryCode: 'GG'),
+      );
 
-    final countryCodeFinder = find.text('+44 1481');
-    final numberFinder = find.text('960194');
+      final countryCodeFinder = find.text('+44 1481');
+      final numberFinder = find.text('960194');
 
-    expect(countryCodeFinder, findsOneWidget);
-    expect(numberFinder, findsOneWidget);
-  });
+      expect(countryCodeFinder, findsOneWidget);
+      expect(numberFinder, findsOneWidget);
+    },
+  );
 
-  testWidgets('Test intl_phone_field setup with UK number: +447891244567', (WidgetTester tester) async {
-    await tester.pumpWidget(const TestWidget(
-      phoneNumber: '+447891244567',
-      countryCode: 'GB',
-    ));
+  testWidgets('Test intl_phone_field setup with UK number: +447891244567', (
+    WidgetTester tester,
+  ) async {
+    await tester.pumpWidget(
+      const TestWidget(phoneNumber: '+447891244567', countryCode: 'GB'),
+    );
 
     final countryCodeFinder = find.text('+44');
     final numberFinder = find.text('7891244567');